refactor: move PaymentRequiredError to payments_py.x402.types
The langsmith middleware was raising its own internal _X402Rejection
class for the same purpose the langchain decorator's PaymentRequiredError
already serves: signal "payment verification failed, hand the buyer the
402 envelope". Two parallel classes for one concept.

Move PaymentRequiredError to payments_py.x402.types (neutral location,
no extra-specific deps). Both the decorator and the LangSmith Deployment
middleware now import it from there. payments_py.x402.langchain still
re-exports it for backwards compatibility, so existing user code that
does `from payments_py.x402.langchain import PaymentRequiredError`
keeps working unchanged.

Why neutral location: PaymentRequiredError previously lived in
payments_py.x402.langchain.decorator. Importing it required the
[langchain] extra at import time (langchain_core is a hard dep of the
decorator module). [langsmith]-only installations would fail. types.py
is part of the core x402 module - importable from any extra.

51 langsmith middleware tests + 519 full unit suite green.
